extract helpers to a utilities folder - Seperated local fallback provider and local provider as signer into seperate methods
