+++ src/contexts/web3Data/utilities/index.ts
@@ -1,0 +1,3 @@
+export * from './fallbackProvider';
+export * from './injectedProvider';
+export * from './localProvider';
